requirements.txt
asgiref            3.9.1
build              1.2.2.post1
colorama           0.4.6
Django             5.2.4
gunicorn           23.0.0
importlib_metadata 8.7.0
numpy              2.2.6
opencv-python      4.12.0.88
packaging          25.0
pillow             11.3.0
pip                21.2.3
mysqlclient        2.2.7
pypng              0.20220715.0
pyproject_hooks    1.2.0
qrcode             8.2
setuptools         57.4.0
sqlparse           0.5.3
tomli              2.2.1
typing_extensions  4.14.1
tzdata             2025.2
wheel              0.45.1
zipp               3.23.0